Add files via upload
This program requires a user to input a password. Then it will check the provided input string for having at least one:
- Capital letter
- Integer number
- '$' sign

If the result is correct, the program will tell user that his password is good, and then terminate.

If the given input was considered weak, the program will let the user know that the input is not satisfied, and keep working on string until succeed. Another option to terminate the program is to type 'quit'.